日志打印

+++ src/main/java/com/demo/dao/StudentDAO.java
@@ -2,7 +2,6 @@ package com.demo.dao;
 
 import com.demo.model.Student;
 import com.demo.model.StudentExample;
-import java.math.BigDecimal;
 import java.util.List;
 
 public interface StudentDAO {
@@ -10,15 +9,15 @@ public interface StudentDAO {
 
     int deleteByExample(StudentExample example);
 
-    int deleteByPrimaryKey(BigDecimal id);
+    int deleteByPrimaryKey(Integer id);
 
-    BigDecimal insert(Student record);
+    Integer insert(Student record);
 
-    BigDecimal insertSelective(Student record);
+    Integer insertSelective(Student record);
 
     List<Student> selectByExample(StudentExample example);
 
-    Student selectByPrimaryKey(BigDecimal id);
+    Student selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(Student record, StudentExample example);
